scopusCrawler/requirements.txt
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
boto==2.49.0
certifi==2022.6.15.2
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.2
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
distlib==0.3.6
fake-useragent==0.1.11
filelock==3.8.0
greenlet==1.1.3
gzip-reader==0.1
h11==0.13.0
html5lib==1.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
ipython==7.20.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
jmespath==0.10.0
logging==0.4.9.6
lxml==4.6.3
mysqlclient==1.4.4
olefile==0.46
outcome==1.2.0
parsel==1.6.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==2.5.2
prompt-toolkit==3.0.14
Protego==0.1.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==2.0.1
PyDispatcher==2.0.5
Pygments==2.7.1
PyGObject==3.38.0
PyHamcrest==1.9.0
pyOpenSSL==20.0.1
pysmbc==1.0.23
PySocks==1.7.1
python-dotenv==0.21.0
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
scrapy-random-useragent==0.2
selenium==4.4.3
service-identity==18.1.0
sh==1.14.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.4.41
tqdm==4.64.1
traitlets==5.0.5
trio==0.21.0
trio-websocket==0.9.2
Twisted==20.3.0
urllib3==1.26.5
virtualenv==20.16.5
w3lib==1.22.0
wcwidth==0.1.9
webdriver-manager==3.8.3
webencodings==0.5.1
wsproto==1.2.0
zope.interface==5.4.0
